pyproject.toml
[project]
name = "ragops"
version = "2.0.0"
description = "RAG Ops — Query any codebase with AI. Open-source, Docker-first, BYO API key."
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "openai>=1.0",
    "boto3>=1.34",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "httpx>=0.27.0",
]

[project.scripts]
ragops = "services.cli.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["services*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["services"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
